chore: optimize current user first load
