+++ app/_components/Providers.tsx
@@ -5,21 +5,25 @@ import { SessionProvider } from 'next-auth/react';
 import { ReactNode, useState } from 'react';
 
 import RefreshTokenHandler from '@/components/utils/auth/RefetchTokenHandler';
+import { UserPopulated } from '@/schemas/user/populated.schema';
 
 const AuthProvider = dynamic(() => import('@/context/auth/auth.provider'));
 const CsrfProvider = dynamic(() => import('@/context/csrf/csrf.provider'));
 
 type ProvidersProps = {
 	children: ReactNode;
+	values: {
+		currentUser: UserPopulated | null;
+	},
 };
 
-const Providers = ({ children }: ProvidersProps) => {
+const Providers = ({ children, values }: ProvidersProps) => {
 
 	const [ refetchInterval, setRefetchInterval ] = useState(0);
 
 	return (
 		<SessionProvider refetchInterval={ refetchInterval }>
-			<AuthProvider>
+			<AuthProvider currentUser={ values.currentUser } >
 				<CsrfProvider>
 					{ children }
 					<RefreshTokenHandler setRefetchInterval={ setRefetchInterval } />
