+++ app/api/users/route.ts
@@ -2,7 +2,6 @@ import { parse } from 'url';
 
 import { NextResponse } from 'next/server';
 
-import { renewFileExpiration } from '@/app/_utils/file/renew-file-expiration';
 import { createUser, findUserByEmail, findUsers, findUsersCount } from '@/database/user/user.repository';
 import { connectToDatabase } from '@/lib/database';
 import { AuthenticationProvider } from '@/schemas/authentication-provider';
@@ -12,7 +11,7 @@ import { buildApiError } from '@/utils/api/error';
 import { ApiErrorCode } from '@/utils/api/error/error-codes.util';
 import { ReasonPhrase, StatusCode } from '@/utils/api/http-status';
 import { setServerAuthGuard } from '@/utils/auth';
-import { isFileURLExpired } from '@/utils/file.util';
+import { isFileURLExpired, renewFileExpiration } from '@/utils/file.util';
 import { generatePassword, hashPassword } from '@/utils/password.util';
 
 import { CreateUserSchema } from './_schemas/create-user.schema';
