+++ utils/file.util.ts
@@ -3,6 +3,9 @@ import path from 'path';
 
 import imageCompression from 'browser-image-compression';
 
+import { updateFileURL } from '@/database/file/file.repository';
+import { DEFAULT_URL_EXPIRATION, generateExpirationDate, gitFileSignedURL } from '@/lib/bucket';
+import { connectToDatabase } from '@/lib/database';
 import { IFile } from '@/schemas/file';
 import { ImageMimeType, MimeTypeSchema } from '@/schemas/file/mime-type.schema';
 
@@ -64,4 +67,25 @@ export const isFileURLExpired = (...files: (IFile | null)[]): IFile[] => {
 		}
 		return false;
 	}) as IFile[];
+};
+
+type RenewFileExpirationOptions = {
+	expiration?: number;
+};
+
+export const renewFileExpiration = async (file: IFile | null, options?: RenewFileExpirationOptions): Promise<IFile | null> => {
+	const { expiration = DEFAULT_URL_EXPIRATION.PROFILE_PICTURE } = options || {};
+
+	if (file && file.url_expiration_date && file.url_expiration_date < new Date()) {
+		await connectToDatabase();
+		const photoUrl = await gitFileSignedURL(file.key, expiration);
+		const updatedFile = await updateFileURL({
+			id: file.id,
+			url: photoUrl,
+			url_expiration_date: generateExpirationDate(expiration),
+		});
+		return updatedFile;
+	}
+
+	return file;
 };
